test(infra): move tmux regression tests to dedicated file (lint guard)
The previous commit added two new test functions to
tests/infra/test_devcontainer_attached_mode.py, which is on the
pydoclint exclude list (#938). The check_no_new_funcs_in_pydoclint_excluded.py
guard correctly rejects new defs in excluded files.

Graduating test_devcontainer_attached_mode.py is the project-preferred
resolution, but doing so surfaces a pre-existing DOC101 violation in
test_docker_entrypoint_idle_mode_blocks_attached_mode (which is unrelated
to this PR). Rather than mix lint-cleanup into this fix, move the two new
tests to a dedicated file tests/infra/test_devcontainer_tmux_session.py
with Sphinx-style docstrings that satisfy pydoclint on their own.

Net: same coverage, no scope creep into other tests, exclude list
unchanged.
