+++ tests/infra/test_devcontainer_attached_mode.py
@@ -84,78 +84,6 @@ def test_every_devcontainer_run_args_includes_env_file_attached_mode(
 
 
 @pytest.mark.infra
-def test_every_devcontainer_tmux_profile_uses_independent_session(
-    devcontainer_json_paths: list[Path],
-) -> None:
-    """The VS Code tmux terminal profile must NOT attach to a shared session.
-
-    Regression guard for #1053: `new-session -A -s main` made every terminal
-    attach to the same session named `main`, so opening a second terminal
-    mirrored the first. Each terminal must launch an independent session.
-    """
-    for path in devcontainer_json_paths:
-        config = _load(path)
-        profiles = (
-            config.get("customizations", {})
-            .get("vscode", {})
-            .get("settings", {})
-            .get("terminal.integrated.profiles.linux", {})
-        )
-        tmux_args = profiles.get("tmux", {}).get("args", [])
-        assert "-A" not in tmux_args, (
-            f"{path}: tmux profile must not pass `-A` (attach-if-exists) — "
-            f"causes every VS Code terminal to share one session. Got args: {tmux_args!r}"
-        )
-        assert "main" not in tmux_args, (
-            f"{path}: tmux profile must not target a fixed session name like `main` — "
-            f"causes every VS Code terminal to share one session. Got args: {tmux_args!r}"
-        )
-
-
-@pytest.mark.infra
-def test_post_create_installs_tmux_conf_to_home(
-    post_create_script: Path,
-) -> None:
-    """post-create.sh must install .devcontainer/tmux.conf to ~/.tmux.conf.
-
-    Regression guard for #1053: tmux mouse mode (and other settings) is shipped by installing the
-    conf into the current user's HOME from post-create.sh. The install must run BEFORE the root→dev
-    exec so the root variant's /root/.tmux.conf is populated for terminals that open as root.
-    """
-    script = post_create_script.read_text()
-    install_line = next(
-        (
-            i
-            for i, line in enumerate(script.splitlines(), start=1)
-            if line.startswith("install ") and "tmux.conf" in line and '"$HOME/.tmux.conf"' in line
-        ),
-        None,
-    )
-    assert install_line is not None, (
-        f"{post_create_script}: must contain a line installing tmux.conf to "
-        f'"$HOME/.tmux.conf" (e.g., `install -m 0644 "$_devc_dir/tmux.conf" '
-        f'"$HOME/.tmux.conf"`).'
-    )
-    root_drop_line = next(
-        (
-            i
-            for i, line in enumerate(script.splitlines(), start=1)
-            if line.strip().startswith('if [ "$(id -u)" -eq 0 ]')
-        ),
-        None,
-    )
-    assert root_drop_line is not None, (
-        f"{post_create_script}: expected a root→dev privilege-drop guard "
-        f'(`if [ "$(id -u)" -eq 0 ]; then ...`).'
-    )
-    assert install_line < root_drop_line, (
-        f"{post_create_script}: tmux.conf install (line {install_line}) must run "
-        f"BEFORE the root→dev exec (line {root_drop_line}) so /root/.tmux.conf "
-        f"is populated for the root variant."
-    )
-
-
-@pytest.mark.infra
 def test_docker_entrypoint_idle_mode_blocks_attached_mode(project_root: Path) -> None:
     """docker_entrypoint.py's idle command must exec `sleep infinity` (not exit).
